Fix indentation in deb package description
The preceding commit 6042c72a7f5e38ec9df1b8a41cee2753f0df45bf manually
indented the body of the package description.  This is not necessary, as
commands like `apt show` will handle this by themselves.

The manually-inserted `.` between paragraphs is also not necessary, as
this will be inserted in place of a blank line automatically during the
creation of the debian control file.

See:
https://www.debian.org/doc/debian-policy/ch-controlfields.html#s-f-Description

... for details on Debian control file description fields.  It looks as
if `fpm` takes care of making sure this spec is adhered to for multiple
paragraphs in package descriptions.