Include copyright notice, contributors, and license in the deb package
MXNet includes a separate `NOTICE` file for its copyright declaration,
together with the `LICENSE` file containing the Apache License v2.0.
This patch updates the deb package definition to make sure that both are
included in `/usr/share/doc/libmxnet/`.  Since the copyright declaration
only refers to 'Contributors', the contributor list is also included.

Note that in principle the `copyright` file should have a fairly strict
format [1], but this is left for a future package update, since there is
no strong need at present for this file to be machine-readable.

[1] https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/