Add custom build script to handle deb package versioning
This script has been created to handle version tags that correspond
directly to package versions of the form:

    UPSTREAM_VERSION-PACKAGE_ITERATION

This is intended to allow us to tag our MXNet package versions with well
crafted Debian package versions, e.g.:

    1.0.0-0tsunami1

where `1.0.0` is the upstream MXNet version, and the package iteration
`0tsunami1` follows the `debian_revision` format (where the initial 0
indicates that we are not forking any upstream Debian package, the
`tsunami` identifies Sociomantic Tsunami as the packager, and the last
number corresponds to our own package revision 1, 2, 3, ..., restarting
from 1 each time the upstream version changes).

The script parses the tag and separates out these version and iteration
components, appending dirty-build and distro-code details to the latter
in a manner compatible with Debian version standards.  It then invokes
Makd to build the package with these version specifications.

Travis CI config has been updated to use the script instead of invoking
`make -f makd/Makd.d` directly.

See https://www.debian.org/doc/debian-policy/#version for general info
on Debian package versioning, and
https://askubuntu.com/questions/620533/how-does-ubuntu-name-packages
for some information on the `XnameY` format for the `debian_revision`.

Finally, note that the current script makes no attempt to distinguish
the package `epoch`.  This can be added later if it proves necessary.