Ensure that all library files are included in libmxnet package
The original `.pkg` file hardcoded assumptions about names of libraries
generated by the MXNet build process.  This patch updates things so that
all files in the `./lib` dir will be included in the package.

This should not produce any change in the package contents for now, but
should make the definition more robust to any changes to the libraries
generated by the MXNet build.