Add custom upload rules for deb packages
As of beaver v0.6.0 we have the option to use the `-C` option to specify
the repo to which a package should be uploaded.  This patch adds a small
helper script to decide the repo based on the version tag: anything with
a whole word equal `alpha`, `beta` or `rc` will go to the `prerelease`
repo, while a package built from any other version tag will go to the
`release` repo.  Travis CI config has been updated to use this script
for the deployment stage.

This should allow us to make trial package release tags along the lines
of `X.y.z-0tsunamiN+rc.1`, validate they work, and then follow up with a
final `X.y.z-0tsunamiN` tag to mark the final release package.