Disable F16C instruction set when building on Travis
This change configures the Travis build without using the F16C
instruction set. Though this instruction set should result in better
performance it reduces portability. CPUs may not support the instruction
set. The used `Makefile` detects whether the CPU performing the build
supports F16C. On Travis the build may be executed on a machine
supporting F16C resulting in libraries that won't work on non-supported
CPUs. Here, we explicitly disable the use of the F16C instruction set
when building on Travis to improve portability of the built libraries.